(PC-4937) Moving icons to the svg folder
