+++ src/ui/svg/icons/ArrowPrevious.tsx
