+++ src/ui/svg/icons/Close.tsx
