+++ src/ui/svg/icons/types.ts
